Refactor snapshot loading
