+++ diagram_analysis/run_mode.py
@@ -1,0 +1,10 @@
+"""Mode tag for analyze runs (wire payloads, TypedDicts, CLI emits)."""
+
+from __future__ import annotations
+
+from enum import StrEnum
+
+
+class RunMode(StrEnum):
+    FULL = "full"
+    INCREMENTAL = "incremental"
